Add flag to cases, implement UI and make unread cases editable (closes #376)

+++ app/admin/cases.rb
@@ -2,13 +2,25 @@ require 'aa_customizable_default_actions'
 
 ActiveAdmin.register Case do
 
-  actions :index, :show, :destroy
+  actions :index, :show, :edit, :update, :destroy
   config.clear_action_items! # get rid of the default action items, since we have to handle 'delete' on a case-by-case basis
 
   controller do
     load_and_authorize_resource :except => :index
     def scoped_collection
       end_of_association_chain.accessible_by(current_ability)
+    end
+
+    def update
+      c = Case.find(params[:id])
+      unless c.state == :unread and c.form_answer.nil?
+        flash[:error] = 'You are not authorized to edit this case!'
+        redirect_to :action => :show
+        return
+      end
+      
+      authorize! :manage, c
+      update!
     end
   end
 
@@ -19,6 +31,16 @@ ActiveAdmin.register Case do
     column :patient
     column :images
     column :case_type
+    column :flag do |c|
+      case c.flag
+      when :regular
+        status_tag('Regular', :ok)
+      when :validation
+        status_tag('Validation', :warning)
+      when :reader_testing
+        status_tag('Reader Testing', :warning)
+      end
+    end
     column :state do |c|
       case c.state
       when :unread
@@ -31,7 +53,7 @@ ActiveAdmin.register Case do
     end
    
     customizable_default_actions do |resource|
-      resource.form_answer.nil? ? [] : [:destroy]
+      (resource.state == :unread and resource.form_answer.nil?) ? [] : [:edit, :destroy]
     end
   end
 
@@ -42,6 +64,16 @@ ActiveAdmin.register Case do
       row :patient
       row :images
       row :case_type
+      row :flag do
+        case c.flag
+        when :regular
+          status_tag('Regular', :ok)
+        when :validation
+          status_tag('Validation', :warning)
+        when :reader_testing
+          status_tag('Reader Testing', :warning)
+        end
+      end
       row :state do
         case c.state
         when :unread
@@ -58,9 +90,24 @@ ActiveAdmin.register Case do
     end
   end
 
+  form do |f|
+    f.inputs 'Details' do
+      f.input :patient, :collection => f.object.session.patients
+      f.input :images
+      f.input :position
+      f.input :case_type
+      f.input :flag, :as => :radio, :collection => {'Regular' => :regular, 'Validation' => :validation}
+    end
+    
+    f.buttons
+  end
+
   action_item :only => :show do
     # copied from activeadmin/lib/active_admin/resource/action_items.rb#add_default_action_items
-    if controller.action_methods.include?('destroy') and resource.form_answer.nil?
+    if controller.action_methods.include?('edit') and resource.state == :unread and resource.form_answer.nil?
+      link_to(I18n.t('active_admin.edit_model', :model => active_admin_config.resource_label), edit_resource_path(resource))
+    end
+    if controller.action_methods.include?('destroy') and resource.state == :unread and resource.form_answer.nil?
       link_to(I18n.t('active_admin.delete_model', :model => active_admin_config.resource_label),
               resource_path(resource),
               :method => :delete, :data => {:confirm => I18n.t('active_admin.delete_confirmation')})
